+++ src/node/index.ts
@@ -1,2 +1,0 @@
-export * from "./node-readable-stream.js";
-export * from "./stream-deserializer.js";